% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Converts US dollar into euro}
\usage{
convert_usd_to_eur(value)
}
\arguments{
\item{value}{numeric}
}
\description{
Converts US dollar into euro
}
