% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beadandó.R
\name{usd_rate}
\alias{usd_rate}
\title{Return currency values as set in}
\usage{
usd_rate(value, incurrency = "EUR", outcurrency = "USD")
}
\arguments{
\item{value}{numeric value}

\item{incurrency}{string, default: incurrency = 'EUR'}

\item{outcurrency}{string, default: outcurrency = 'EUR'}
}
\value{
the exchange rate of the given currency
}
\description{
Returns the freshest exchange rates from the European Central Bank
}
